- Changed binary target back to float, since it seems like that is what PyTorch wanted - Converted the accuracy targets to integers, since that is what pytorch lightning wanted

+++ datasets/nih_cxr.py
@@ -96,7 +96,7 @@ class NIHCXRDataset(Dataset):
         label = torch.from_numpy(label.values)
         if self.binary:
             label = (label == 1).any()
-        label = label.long()
+        label = label.float()
 
         return image, label
